Refactoring of Schemas
