+++ src/sdk/python/rtdip_sdk/pipelines/destinations/__init__.py
@@ -14,3 +14,4 @@
 
 from .spark.delta import *
 from .spark.eventhub import *
+from .spark.kafka import *
