+++ src/sdk/python/rtdip_sdk/pipelines/sources/spark/kafka.py
@@ -18,7 +18,8 @@ from pyspark.sql import DataFrame, SparkSession
 
 from ..interfaces import SourceInterface
 from ..._pipeline_utils.models import Libraries, SystemType
-from ..._pipeline_utils.constants import DEFAULT_PACKAGES, KAFKA_SCHEMA
+from ..._pipeline_utils.constants import DEFAULT_PACKAGES
+from ..._pipeline_utils.spark import KAFKA_SCHEMA
 
 class SparkKafkaSource(SourceInterface):
     '''
