+++ tests/sdk/python/rtdip_sdk/pipelines/sources/spark/test_kafka.py
@@ -17,7 +17,7 @@ sys.path.insert(0, '.')
 import pytest
 from src.sdk.python.rtdip_sdk.pipelines.sources.spark.kafka import SparkKafkaSource
 from src.sdk.python.rtdip_sdk.pipelines._pipeline_utils.models import Libraries, MavenLibrary
-from src.sdk.python.rtdip_sdk.pipelines._pipeline_utils.constants import KAFKA_SCHEMA
+from src.sdk.python.rtdip_sdk.pipelines._pipeline_utils.spark import KAFKA_SCHEMA
 from tests.sdk.python.rtdip_sdk.pipelines._pipeline_utils.spark_configuration_constants import spark_session
 import json
 from pyspark.sql import DataFrame, SparkSession
